Pre tests

+++ Project10/ColorTable.cpp
@@ -57,9 +57,9 @@ void ColorTable::insertGradient(const Color& color1, const Color& color2, const 
 	if (position2 > position1 && position1 >= 0 && position1 < getNumberOfColors() && position2 >= 0 && position2 < getNumberOfColors())
 	{
 		int pos_diff = position2 - position1;
-		int red_delta = (color2.getRed() - color1.getRed()) / pos_diff;
-		int green_delta  = (color2.getGreen() - color1.getGreen()) / pos_diff;
-		int blue_delta = (color2.getBlue() - color1.getBlue()) / pos_diff;
+		double red_delta = (color2.getRed() - color1.getRed()) / pos_diff;
+		double green_delta  = (color2.getGreen() - color1.getGreen()) / pos_diff;
+		double blue_delta = (color2.getBlue() - color1.getBlue()) / pos_diff;
 
 		for (int i = 0; i <= pos_diff; i++)
 		{
